01-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
aiofiles # Necessário para servir arquivos estáticos como o frontend